pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "atp-protocol"
version = "1.0.0"
description = "ATP Protocol - Agentic Trade Protocol. The easiest way to enable agent-to-agent payments powered by Solana."
license = "MIT"
authors = ["The Swarm Corporation"]
homepage = "https://github.com/The-Swarm-Corporation/ATP-Protocol"
documentation = "https://github.com/The-Swarm-Corporation/ATP-Protocol"
readme = "README.md"
repository = "https://github.com/The-Swarm-Corporation/ATP-Protocol"
packages = [{include = "atp"}]
keywords = [
    "atp",
    "atp-protocol",
    "solana",
    "blockchain",
    "cryptocurrency",
    "payment-gated",
    "agent-to-agent",
    "agent payments",
    "agent execution",
    "payment settlement",
    "solana payments",
    "usdc",
    "sol",
    "fastapi",
    "agent api",
    "payment gateway",
    "settlement service",
    "agent marketplace",
    "decentralized payments",
    "web3",
    "crypto payments",
    "agent infrastructure",
    "payment middleware",
    "automated settlement",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Office/Business :: Financial",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
setuptools = "*"
httpx = "*"
loguru = "*"
pydantic = "*"
solana = "*"
solders = "*"
python-dotenv = "*"


[tool.poetry.group.lint.dependencies]
black = ">=23.1,<26.0"
ruff = ">=0.5.1,<0.14.11"
types-toml = "^0.10.8.1"
types-pytz = ">=2023.3,<2026.0"
types-chardet = "^5.0.4.6"
mypy-protobuf = ">=3,<5"


[tool.poetry.group.test.dependencies]
pytest = ">=8.1.1,<10.0.0"

[tool.poetry.group.dev.dependencies]
black = "*"
ruff = "*"
pytest = "*"
# pre-commit = "*"

[tool.ruff]
line-length = 70

[tool.black]
target-version = ["py38"]
line-length = 70
include = '\.pyi?$ | \.rs$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''

